refactor: inject 메서드명을 명시적으로 변경
