+++ app/src/main/java/woowacourse/shopping/ui/serversetting/ServerSettingActivity.kt
@@ -6,13 +6,15 @@ import woowacourse.shopping.databinding.ActivityServerSettingBinding
 import woowacourse.shopping.ui.serversetting.ServerSettingContract.Presenter
 import woowacourse.shopping.ui.shopping.ShoppingActivity
 import woowacourse.shopping.util.extension.setContentView
-import woowacourse.shopping.util.inject.inject
+import woowacourse.shopping.util.inject.injectServerSettingPresenter
 import woowacourse.shopping.util.preference.ShoppingPreference
 
 
 class ServerSettingActivity : AppCompatActivity(), ServerSettingContract.View {
     private lateinit var binding: ActivityServerSettingBinding
-    private val presenter: Presenter by lazy { inject(this, ShoppingPreference(this)) }
+    private val presenter: Presenter by lazy {
+        injectServerSettingPresenter(view = this, shoppingPreference = ShoppingPreference(this))
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
