+++ app/src/main/java/woowacourse/shopping/util/inject/DatabaseInject.kt
@@ -3,6 +3,6 @@ package woowacourse.shopping.util.inject
 import android.content.Context
 import woowacourse.shopping.data.database.ShoppingDatabase
 
-fun createShoppingDatabase(context: Context): ShoppingDatabase {
+fun injectShoppingDatabase(context: Context): ShoppingDatabase {
     return ShoppingDatabase.get(context)
 }
